second version

+++ app.py
@@ -11,9 +11,17 @@ app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 ALLOWED_EXTENSIONS = set(['png,jpg'])
 app.config['MAX_CONTENT_LENGTH'] = 16 * 1024 * 1024
 MAX_FILE_SIZE = 10240 * 1024 + 1
-uploads_dir = os.path.join(app.instance_path, 'uploads')
+uploads_dir = app.static_folder
 
 global name
+
+
+def allowed_file(filename):
+    s = str(filename)
+    if (s[-3:] == 'jpg') or (s[-3:] == 'png'):
+        return True
+    else:
+        return False
 
 
 @app.route('/upload', methods=['GET', 'POST'])
@@ -21,18 +29,23 @@ def upload():
     if request.method == "POST":
         names = os.listdir('static')
         for i in names:
-            os.remove(f'static/{i}')
+            if (os.path.isfile(i)) and (i != 'picture.jpg'):
+                os.remove(f'static/{i}')
         file = request.files["file"]
-        file.save(os.path.join(uploads_dir, 'picture.jpg'))
-        name = str(file.filename)
-        seg(name)
+        print(file.filename)
+        print(allowed_file(file.filename))
+        if allowed_file(file.filename):
+            file.save(os.path.join(uploads_dir, 'picture.jpg'))
+            name = 'picture.jpg'
+            print('пересохранение изображения')
+            seg('picture.jpg')
         return redirect('/result')
     return render_template("loading.html")
 
 
 @app.route('/result')
 def result():
-    return render_template('result.html', name=name)
+    return render_template('result.html')
 
 
 @app.route('/')
